+++ src/FreeElephants/TSxUnit/Suite/Summary.ts
@@ -2,8 +2,48 @@ namespace FreeElephants.TSxUnit.Suite {
 
     export class Summary {
 
-        public constructor(passedList: ResultList, failedList: ResultList, errorList: ResultList) {
+        private passedList: ResultList;
+        private failedList: ResultList;
+        private errorList: ResultList;
+        private totalNumberOfAssertions: number;
 
+        public constructor(passedList: ResultList, failedList: ResultList, errorList: ResultList, totalNumberOfAssertions: number) {
+            this.passedList = passedList;
+            this.failedList = failedList;
+            this.errorList = errorList;
+            this.totalNumberOfAssertions = totalNumberOfAssertions;
+        }
+
+        public getNumberOfPassed(): number {
+            return this.getPassedList().count();
+        }
+
+        public getNumberOfErrors(): number {
+            return this.getErrorList().count();
+        }
+
+        public getNumberOfFailed(): number {
+            return this.getFailedList().count();
+        }
+
+        public getErrorList(): FreeElephants.TSxUnit.Suite.ResultList {
+            return this.errorList;
+        }
+
+        public getFailedList(): FreeElephants.TSxUnit.Suite.ResultList {
+            return this.failedList;
+        }
+
+        public getPassedList(): FreeElephants.TSxUnit.Suite.ResultList {
+            return this.passedList;
+        }
+
+        public isOk(): boolean {
+            return this.getNumberOfErrors() === 0 && this.getNumberOfFailed() === 0;
+        }
+
+        public getNumberOfAssertions(): number {
+            return this.totalNumberOfAssertions;
         }
     }
 }