+++ src/reference.ext.ts
@@ -8,4 +8,3 @@
 ///<reference path="FreeElephants/TSxUnit/Suite/Summary.ts"/>
 ///<reference path="FreeElephants/TSxUnit/Suite/ResultList.ts"/>
 
-///<reference path="FreeElephants/TSxUnit/Test/TestCaseMethod.ts"/>